requirements.txt
requests>=2.25.1
pyvis>=0.3.1
PyQt5>=5.15.0
python-dotenv>=0.19.0
qt-material>=2.14